requirements.txt
# FastAPI and web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.1

# PDF manipulation
pypdf>=4.2.0

# Dropbox integration
dropbox>=12.0.0

# Google Sheets integration
google-auth>=2.29.0
google-api-python-client>=2.125.0

# HTTP client
httpx>=0.27.0
